archive/requirements_tf23.txt
lightgbm==3.2.1
matplotlib==3.0.3
numpy==1.18.5
opencv-python==4.4.0.46
openpyxl==3.0.5
pandas==1.1.5
pandas_streaming==0.2.175
Pillow==8.0.1
psutil==5.8.0
scikit-image==0.17.1
scikit-learn==0.24.0
seaborn==0.11.1
tensorflow-gpu==2.3.1
